+++ __tests__/useCkoSofort.spec.ts
@@ -94,6 +94,27 @@ describe('[checkout-com] useCkoSofort', () => {
 
   });
 
+  it('calls createPayment & returns proper success response & clears error if any from previous try', async () => {
+
+    /*eslint-disable */
+    const payload = {
+      cartId: 15,
+      contextDataId: 'abc',
+      email: 'ab@gmail.com',
+      success_url: null,
+      failure_url: null
+    };
+    /* eslint-enable */
+
+    error.value = 'mockd';
+    const response = await makePayment(payload);
+
+    expect(createPayment).toHaveBeenCalled();
+    expect(response).toEqual(defaultPaymentResponse);
+    expect(error.value).toBe(null);
+
+  });
+
   it('uses default values for success and failure url and reference', async () => {
 
     /*eslint-disable */