+++ src/useCkoSofort.ts
@@ -34,6 +34,7 @@ const useCkoSofort = () => {
         throw new Error(payment.data.error_type);
       }
 
+      error.value = null;
       return payment;
     } catch (e) {
       error.value = e;