+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.9 (not released)
 - Sofort support ([#5158](https://github.com/DivanteLtd/vue-storefront/issues/5158))
 - Klarna support ([#5157](https://github.com/DivanteLtd/vue-storefront/pull/5157))
+- Clearing error private refs after succesful payment & error public ref before makePayment action from payment method ([#6](https://github.com/vuestorefront/checkout-com/pull/6))
 
 ## 0.0.8